Add support to change the base URL of service.
This change add the getter and setter method to make available the path of bitcoin node to outside.

This is particular useful when the user want work with different wallet in the same node.

The unit test to prof that this change works is available here: https://gitlab.com/vincenzopalazzo/draft-ideas/-/blob/master/pyln-test-sandbox/test/test_multiple_wallet.py

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>